+++ front/src/app/app.routing.ts
@@ -21,6 +21,8 @@ import {ModuleEditorComponent} from './itinerary/module/moduleEditor/module-edit
 import {CourseComponent} from './course/course.component';
 import {ModuleFormComponent} from './itinerary/module/moduleForm/module-form.component';
 import {UnitsBlocksToolComponent} from './itinerary/module/moduleEditor/units-blocks-tool.component';
+import {ImageComponent} from './images/image.component';
+import {LessonSlidesToolComponent} from './itinerary/lesson/lessonTools/lesson-slides-tool.component';
 import {QuestionTrackingComponent} from './question/questionTracking/questionTracking.component';
 
 
@@ -29,6 +31,7 @@ const appRoutes = [
   { path: 'unit' , component: ViewComponent },
   { path: 'unit/:unitId', component: UnitComponent },
   { path: 'unitsCardsTool', component: UnitsCardsToolComponent },
+  { path: 'lessonSlidesTool', component: LessonSlidesToolComponent },
   { path: 'unitsBlocksTool', component: UnitsBlocksToolComponent },
   { path: 'lessonForm', component: LessonFormComponent },
   { path: 'moduleForm', component: ModuleFormComponent },
@@ -50,7 +53,8 @@ const appRoutes = [
   { path: 'courses', component: MyCoursesComponent },
   { path: 'progress', component: ModuleProgressComponent },
   { path: 'newCourse', component: NewCourseComponent},
-  { path: 'course/:courseId', component: CourseComponent}
+  { path: 'course/:courseId', component: CourseComponent},
+  { path: 'images', component: ImageComponent}
 
 ];