+++ front/src/app/card/card.component.ts
@@ -6,6 +6,16 @@ import {Card} from './card.model';
 import {UnitService} from '../unit/unit.service';
 import {MatDialog} from '@angular/material/dialog';
 import {ConfirmActionComponent} from '../confirmAction/confirm-action.component';
+import Asciidoctor from 'asciidoctor';
+import {ImageComponent} from '../images/image.component';
+import {MatBottomSheet} from '@angular/material/bottom-sheet';
+import {ImageService} from '../images/image.service';
+
+function convertToHTML(text) {
+  const asciidoctor = Asciidoctor();
+  const html = asciidoctor.convert(text);
+  return(html);
+}
 
 @Component({
   selector: 'app-cards',
@@ -22,20 +32,37 @@ export class CardComponent implements OnInit, AfterViewChecked {
   unit: Unit;
   showSpinner = false;
 
+  contentCount: number;
+
+  cardsView: boolean[];
+  prettyCards: any[];
+
   confirmText = 'Se eliminará la ficha permanentemente';
   button1 = 'Cancelar';
   button2 = 'Borrar';
 
-  constructor(private router: Router, private activatedRoute: ActivatedRoute, private cardService: CardService,
-              private unitService: UnitService, private dialog: MatDialog) {}
+  constructor(private router: Router,
+              private activatedRoute: ActivatedRoute,
+              private cardService: CardService,
+              private unitService: UnitService,
+              private dialog: MatDialog,
+              private bottomSheet: MatBottomSheet,
+              private imageService: ImageService) { }
 
   ngOnInit() {
     this.activatedRoute.params.subscribe(params => {
       this.unitId = params.unitId;
       this.showSpinner = true;
+      this.cardsView = [];
+      this.prettyCards = [];
       this.unitService.getUnit(this.unitId).subscribe((data: Unit) => {
         this.unit = data;
         this.cards = this.unit.cards;
+        this.cards.forEach((card, index) => {
+          this.cardsView.push(true);
+          this.prettyCards.push('');
+          this.cardContentProcessor(card.content, index);
+        });
         this.showSpinner = false;
       });
     });
@@ -49,6 +76,96 @@ export class CardComponent implements OnInit, AfterViewChecked {
     });
   }
 
+  cardContentProcessor(content: string, index: number) {
+    this.contentCounterFunction(content);
+    let extractedData = [];
+    let position = [];
+    let counter = 0;
+    let contentCounter = 0;
+    const lineas = content.split('\n');
+    lineas.forEach((line) => {
+      const words = line.split('.');
+      if (words[0] === 'insert') {
+        const parameters = words[1].split('/');
+        if (parameters[0] === 'image') {
+          position.push(counter);
+          this.getEmbebedContent(Number(parameters[1]), content, contentCounter, 'image', extractedData, position, index);
+          contentCounter = contentCounter + 1;
+        }
+      }
+      this.addExtractedData(content, extractedData, position, index);
+      counter = counter + 1;
+    });
+  }
+
+  addExtractedData(content: string, extractedData: any, position: any, index: number) {
+    let componentsChecker = 0;
+    let cardContentExtended = '';
+    let lines = content.split('\n');
+    for (let i = 0; i < position.length; i ++) {
+      lines[position[i]] = extractedData[i];
+    }
+    lines.forEach((line: string) => {
+      cardContentExtended = cardContentExtended + line + '\n';
+    });
+    extractedData.forEach((component: string) => {
+      if (component !== '') {
+        componentsChecker = componentsChecker + 1;
+      }
+    });
+    if (componentsChecker === this.contentCount) {
+      this.showSpinner = false;
+      this.prettyCards[index] = convertToHTML(cardContentExtended);
+    }
+  }
+
+  async getEmbebedContent(content1: any, content: string, contentCounter: number, type: string, extractedData: any, position: any, index: number) {
+    let contentEmbebed;
+
+    if (type === 'image') {
+      contentEmbebed = await this.imageService.getImage(content1).toPromise().catch((error) => console.log(error));
+      if (typeof contentEmbebed !== 'undefined') {
+        const image = this.convertImage(contentEmbebed.image);
+        const img = '++++\n' +
+          '<img class="img-lesson" src="' + image + '">\n' +
+          '++++\n' +
+          '\n';
+        extractedData.splice(contentCounter, 1, img);
+      } else {
+        extractedData.splice(contentCounter, 1, 'ERROR\n');
+      }
+    }
+
+    this.addExtractedData(content, extractedData, position, index);
+  }
+
+  contentCounterFunction(content: string) {
+    this.contentCount = 0;
+    let lines: string[];
+    lines = content.split('\n');
+    lines.forEach((line: string) => {
+      let words: string[];
+      words = line.split('.');
+      if (words[0] === 'insert') {
+        this.contentCount = this.contentCount + 1;
+      }
+    });
+  }
+
+  convertImage(bytes: any) {
+    return 'data:image/png;base64,' + btoa(new Uint8Array(bytes).reduce((data, byte) =>
+      data + String.fromCharCode(byte),
+      ''));
+  }
+
+  editCard(index: number) {
+    this.cardsView[index] = false;
+  }
+
+  viewCard(index: number) {
+    this.cardsView[index] = true;
+  }
+
   private addCard() {
     if ((this.cards.length === 0) || (this.cards[0].id !== 0) || (this.cards[0].name && this.cards[0].content)) {
       this.cards.unshift({
@@ -56,6 +173,8 @@ export class CardComponent implements OnInit, AfterViewChecked {
         name: '',
         content: ''
       });
+      this.prettyCards.unshift(convertToHTML(''));
+      this.cardsView.unshift(false);
     }
     this.focusNewCard();
   }
@@ -74,7 +193,7 @@ export class CardComponent implements OnInit, AfterViewChecked {
     return index;
   }
 
-  private deleteCard(id: number) {
+  private deleteCard(id: number, index: number) {
     const dialogRef = this.dialog.open(ConfirmActionComponent, {
       data: {confirmText: this.confirmText, button1: this.button1, button2: this.button2}
     });
@@ -84,10 +203,14 @@ export class CardComponent implements OnInit, AfterViewChecked {
         this.showSpinner = true;
         if (id === 0) {
           this.cards.splice(0, 1);
+          this.prettyCards.splice(0, 1);
+          this.cardsView.splice(0, 1);
           this.showSpinner = false;
         } else {
           this.cardService.deleteCard(this.unitId, id).subscribe(() => {
             this.cards.splice(this.getCardIndex(id), 1);
+            this.prettyCards.splice(index, 1);
+            this.cardsView.splice(index, 1);
             this.showSpinner = false;
           });
         }
@@ -95,16 +218,18 @@ export class CardComponent implements OnInit, AfterViewChecked {
     });
   }
 
-  save(card: Card) {
+  save(card: Card, index: number) {
     if (card.name && card.content) {
       this.showSpinner = true;
       if (card.id === 0) {
         this.cardService.create(this.unitId, card).subscribe((createdCard) => {
           card.id = createdCard.id;
+          this.cardContentProcessor(card.content, index);
           this.showSpinner = false;
         });
       } else {
         this.cardService.save(this.unitId, card).subscribe(() => {
+          this.cardContentProcessor(card.content, index);
           this.showSpinner = false;
         }, error => {
           console.error(error);
@@ -121,6 +246,10 @@ export class CardComponent implements OnInit, AfterViewChecked {
     textArea.style.overflow = 'hidden';
     textArea.style.height = '0px';
     textArea.style.height = textArea.scrollHeight + 'px';
+  }
+
+  openImageBottomSheet(): void {
+    this.bottomSheet.open(ImageComponent);
   }
 
 }